Extract People into it's own file
